Full CRUD basic

+++ packages/backend/src/modules/task-manager/task-manager.service.ts
@@ -1,9 +1,18 @@
+import { Repository } from 'typeorm';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { TaskEntity } from 'src/entities/task.entity';
-import { Repository } from 'typeorm';
-import { AddTaskDto, TasksResponseDto } from './task-manager.dto';
+import {
+  InvalidTeamException,
+  NotFoundTaskException,
+} from '@core/exceptions/task';
+import { TaskEntity, TeamType } from 'src/entities/task.entity';
+
+import {
+  AddTaskDto,
+  TasksResponseDto,
+  UpdateTaskDto,
+} from './task-manager.dto';
 
 @Injectable()
 export class TaskManagerService {
@@ -20,14 +29,42 @@ export class TaskManagerService {
     };
   }
 
-  async createTask(task: AddTaskDto) {
+  async createTask(input: AddTaskDto) {
     return this.taskRepo.save({
-      team: task.team,
-      description: task.description,
+      team: input.team,
+      description: input.description,
     });
   }
 
+  async updateTask(input: UpdateTaskDto) {
+    const task = await this.taskRepo.findOneBy({ id: input.id });
+
+    if (!task) {
+      throw new NotFoundTaskException();
+    }
+
+    if (input.team && !Object.values(TeamType).includes(input.team)) {
+      throw new InvalidTeamException();
+    }
+
+    await this.taskRepo
+      .createQueryBuilder('task')
+      .update(TaskEntity)
+      .when(input.team, (query) => query.set({ team: input.team }))
+      .when(input.description, (query) =>
+        query.set({ description: input.description }),
+      )
+      .where('id = :id', { id: input.id })
+      .execute();
+  }
+
   async deleteTask(id: number) {
+    const task = await this.taskRepo.findOneBy({ id });
+
+    if (!task) {
+      throw new NotFoundTaskException();
+    }
+
     return this.taskRepo.delete({ id });
   }
 }
